+++ examples/partitioned-store-restapi/src/main/java/dev/thriving/oss/kafka/streams/cassandra/state/store/example/partitionedstore/restapi/KTablePartitionedStoreRestApiDemo.java
@@ -228,7 +228,7 @@ public final class KTablePartitionedStoreRestApiDemo {
     }
 
     /**
-     * Get the optimised special implementation of ReadOnlyKeyValueStore (CassandraReadOnlyKeyValueStore).
+     * Get the optimised special implementation of ReadOnlyKeyValueStore (CassandraPartitionedReadOnlyKeyValueStore).
      * Lazy init, re-use.
      * @return instance of ReadOnlyKeyValueStore
      */