+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Offered by `CassandraStateStore` static methods, for an example please check the [README](README.md#interactive-queries)
     - `CassandraStateStore#readOnlyPartitionedKeyValueStore(KafkaStreams, String, CqlSession, String, boolean, String, Serde, Serde)`
     - `CassandraStateStore#readOnlyPartitionedKeyValueStore(KafkaStreams, String, CqlSession, String, boolean, String, Serde, Serde, Function, StreamPartitioner)`
-- New example '[processor-api-all-range-prefix-count](examples/processor-api-all-range-prefix-count)' featuring the new `CassandraReadOnlyKeyValueStore`
+- New example '[processor-api-all-range-prefix-count](examples/processor-api-all-range-prefix-count)' featuring the new `CassandraPartitionedReadOnlyKeyValueStore`
   - Implements a REST API with endpoints for all available 'Interactive Query' methods to play with hands-on
 
 ### Changed
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - test + assertions added for `ReadOnlyKeyValueStore.all()`
 
 ### Removed
-- First impl. of `CassandraStateStore#readOnlyPartitionedKeyValueStore(KafkaStreams, String)` was removed and replaced by the more complex methods that return an instance of `CassandraReadOnlyKeyValueStore` (already mentioned above) 
+- First impl. of `CassandraStateStore#readOnlyPartitionedKeyValueStore(KafkaStreams, String)` was removed and replaced by the more complex methods that return an instance of `CassandraPartitionedReadOnlyKeyValueStore` (already mentioned above) 
 
 
 ## [0.6.0](https://github.com/thriving-dev/kafka-streams-cassandra-state-store/releases/tag/0.6.0) (2023-07-11)